+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - flake8 3.7.4 -> 3.7.5
 - twine 1.12.1 -> 1.13.0
 - fixed bug in ```run_shellcheck.sh``` where docker containers weren't being removed are they had exited
+- using ```"$(echo "${1:-}" | tr "[:upper:]" "[:lower:]")"``` instead of ```"${1,,}"``` so scripts work on macOS and Ubuntu
 
 ### Removed